requirements.txt
# Core Dependencies
yt-dlp>=2023.0.0
whisperx>=3.1.1
openai-whisper>=20230918
google-genai>=0.3.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
PyYAML>=6.0
python-dotenv>=1.0.0

# GPU Support (Optional but Recommended)
# torch>=2.0.0
# torchaudio>=2.0.0
# torchvision>=0.15.0
